IrregularEventViewController was fully supplemented with UICollectionViews for color and emoji. DataSourse and DelegateFlowLayout were added and set. According Cells and Headers were added and set up.
