+++ Tracker/TrackersViewController/AddingTracker/NewTrackerViewController/IrregularEvent/EventColorHeader.swift
@@ -1,0 +1,41 @@
+//
+//  EventColorHeader.swift
+//  Tracker
+//
+//  Created by Almira Khafizova on 23.11.23.
+//
+
+import UIKit
+
+final class EventColorHeader: UICollectionReusableView {
+    
+    static let id = "header"
+    var headerText: String? {
+        didSet {
+            titleLabel.text = headerText
+        }
+    }
+    
+    private lazy var titleLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .ypBlackDay
+        label.font = UIFont.systemFont(ofSize: 19, weight: .bold)
+        label.translatesAutoresizingMaskIntoConstraints = false
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        addSubview(titleLabel)
+
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 0),
+            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10)
+        ])
+    }
+    
+    required init?(coder: NSCoder) {
+        assertionFailure("init(coder:) has not been implemented")
+        return nil
+    }
+}
