+++ Tracker/TrackersViewController/AddingTracker/NewTrackerViewController/IrregularEvent/EventEmojiCell.swift
@@ -1,14 +1,15 @@
 //
-//  HabitEmojiCell.swift
+//  EventEmojiCell.swift
 //  Tracker
 //
-//  Created by Almira Khafizova on 21.11.23.
+//  Created by Almira Khafizova on 23.11.23.
 //
 
 import UIKit
 
-final class HabitEmojiCell: UICollectionViewCell {
-    static var reuseId = "HabitEmojiCell"
+final class EventEmojiCell: UICollectionViewCell {
+    
+    static var reuseId = "Event emoji cell"
     
     let emojiLabel: UILabel = {
         let label = UILabel()
