+++ Tracker/TrackersViewController/AddingTracker/NewTrackerViewController/RegularHabit/HabitColorCell.swift
@@ -18,16 +18,18 @@ final class HabitColorCell: UICollectionViewCell {
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-
+        
         contentView.addSubview(colorView)
     }
     override func layoutSubviews() {
         super.layoutSubviews()
-
-        colorView.frame = CGRect(x: (contentView.bounds.width - 40) / 2,
-                                 y: (contentView.bounds.height - 40) / 2,
-                                 width: 40,
-                                 height: 40)
+        
+        colorView.frame = CGRect(
+            x: (contentView.bounds.width - 40) / 2,
+            y: (contentView.bounds.height - 40) / 2,
+            width: 40,
+            height: 40
+        )
     }
     
     required init?(coder: NSCoder) {
