+++ Tracker/TrackersViewController/AddingTracker/NewTrackerViewController/RegularHabit/HabitEmojiCell.swift
@@ -1,0 +1,36 @@
+//
+//  HabitEmojiCell.swift
+//  Tracker
+//
+//  Created by Almira Khafizova on 21.11.23.
+//
+
+import UIKit
+
+final class HabitEmojiCell: UICollectionViewCell {
+    static var reuseId = "HabitEmojiCell"
+    
+    let emojiLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = UIFont.systemFont(ofSize: 32)
+        label.textAlignment = .center
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        contentView.addSubview(emojiLabel)
+        
+        NSLayoutConstraint.activate([
+            emojiLabel.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
+            emojiLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor)
+        ])
+    }
+    
+    required init?(coder: NSCoder) {
+        assertionFailure("init(coder:) has not been implemented")
+        return nil
+    }
+}
