+++ Tracker/TrackersViewController/AddingTracker/NewTrackerViewController/RegularHabit/HabitEmojiHeader.swift
@@ -26,7 +26,7 @@ final class HabitEmojiHeader: UICollectionReusableView {
     override init(frame: CGRect) {
         super.init(frame: frame)
         addSubview(titleLabel)
-
+        
         NSLayoutConstraint.activate([
             titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 0),
             titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10)
